app/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
django_google_maps
django-smart-selects
django-colorfield
googlemaps
django-json-widget
django-import-export
haversine
django-autocomplete-light==3.5.1
django-sslserver
django-mptt
Pillow
django-quill-editor
django-taggit
django-crispy-forms
django-filter
django-paypal
django-bootstrap-breadcrumbs
django-multiselectfield
